Pieces of Authorization

+++ controllers/admin.js
@@ -10,9 +10,8 @@ exports.getAddProduct = (req,res,next) => {
 
 exports.postAddproduct = (req,res,next) => {
     const {title,price,description,imageUrl} = req.body;    
-    let product;
-   
-        product = new Product({
+
+    let product = new Product({
             title:title,
             price:price,
             description:description,
@@ -77,10 +76,10 @@ exports.postEditProducts = (req,res) => {
     const {productId,title,price,description,imageUrl} = req.body;                         
 
     const obj = {
-        title: title, 
-        price: price, 
-        description: description, 
-        imageUrl: imageUrl,
+        title, 
+        price,
+        description, 
+        imageUrl,
         userId: req.user._id
     }
 
@@ -98,7 +97,7 @@ exports.postEditProducts = (req,res) => {
         //     product.imageUrl = imageUrl;
         //     return product.save();
         // })
-    Product.updateOne({_id: productId},{ $set: obj },{ upsert:true })
+    Product.updateOne({_id: productId, userId:req.user._id},{ $set: obj },{ upsert:true })
     .then(result => res.redirect(301,'/admin/products'))
     .catch( err => {
         console.log(err);
@@ -108,7 +107,7 @@ exports.postEditProducts = (req,res) => {
 
 exports.postDeleteProducts = (req,res) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId})
+    Product.deleteOne({_id: prodId, userId:req.user._id})
     .then(result => res.redirect(301,'/admin/products'))
     .catch( err => {
         console.log(err);
